+++ src/django/api/views/fields/nonstandard_fields_view_set.py
@@ -4,7 +4,7 @@ from rest_framework.response import Response
 from rest_framework.status import HTTP_401_UNAUTHORIZED
 from rest_framework.viewsets import GenericViewSet
 
-from ...models.nonstandart_field import NonstandardField
+from ...models.nonstandard_field import NonstandardField
 from ..contributor.get_contributor import get_contributor
